New agent, new agent action and other features
In AgentPerception is now new variable named hint. This is hint
to agent from simulator/user. Agent can read this hint and act
based on it (or ignore it).

New AgentActions.NONE -- for no action.

Newly is possible to load agent'c class based on configuration
file. In config file, there should be "agent=AgentName", where
AgentName is agent's class we want to run.